Decouple Tasks from Jobs
